requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
pydantic
python-dotenv
